+++ campaign/campaign_sos/campaign_3_5.py
@@ -69,7 +69,13 @@ class Campaign(CampaignBase):
     MAP = MAP
 
     def battle_0(self):
-        if self.clear_enemy(scale=(2, 3), genre=['light', 'carrier', 'enemy', 'main']):
+        if self.clear_filter_enemy(self.ENEMY_FILTER, preserve=1):
+            return True
+
+        return self.battle_default()
+
+    def battle_2(self):
+        if self.clear_filter_enemy(self.ENEMY_FILTER, preserve=0):
             return True
 
         return self.battle_default()
