+++ campaign/campaign_sos/campaign_4_5.py
@@ -79,7 +79,13 @@ class Campaign(CampaignBase):
     MAP_ENEMY_SEARCHING_OVERLAY_TRANSPARENCY_THRESHOLD = 0.65
 
     def battle_0(self):
-        if self.clear_enemy(scale=(2, 3), genre=['light', 'carrier', 'enemy', 'main']):
+        if self.clear_filter_enemy(self.ENEMY_FILTER, preserve=1):
+            return True
+
+        return self.battle_default()
+
+    def battle_3(self):
+        if self.clear_filter_enemy(self.ENEMY_FILTER, preserve=0):
             return True
 
         return self.battle_default()
