+++ campaign/campaign_sos/campaign_5_5.py
@@ -73,7 +73,13 @@ class Campaign(CampaignBase):
     MAP = MAP
 
     def battle_0(self):
-        if self.clear_enemy(scale=(2, 3), genre=['light', 'carrier', 'enemy', 'main']):
+        if self.clear_filter_enemy(self.ENEMY_FILTER, preserve=1):
+            return True
+
+        return self.battle_default()
+
+    def battle_3(self):
+        if self.clear_filter_enemy(self.ENEMY_FILTER, preserve=0):
             return True
 
         return self.battle_default()
