+++ campaign/campaign_sos/campaign_8_5.py
@@ -75,10 +75,13 @@ class Campaign(CampaignBase):
     MAP = MAP
 
     def battle_0(self):
-        if self.fleet_2_push_forward():
+        if self.clear_filter_enemy(self.ENEMY_FILTER, preserve=1):
             return True
 
-        if self.clear_enemy(scale=(2, 3), genre=['light', 'carrier', 'enemy', 'main']):
+        return self.battle_default()
+
+    def battle_3(self):
+        if self.clear_filter_enemy(self.ENEMY_FILTER, preserve=0):
             return True
 
         return self.battle_default()
