+++ module/config/argument/menu.json
@@ -15,9 +15,9 @@
       "EventGeneral",
       "Event",
       "Event2",
-      "EventAb",
+      "EventA",
       "EventB",
-      "EventCd",
+      "EventC",
       "EventD",
       "EventSp",
       "Raid",
