+++ module/config/config_updater.py
@@ -362,15 +362,15 @@ class ConfigGenerator:
                     else:
                         insert('Event')
                         insert('Event2')
-                        insert('EventAb')
+                        insert('EventA')
                         insert('EventB')
-                        insert('EventCd')
+                        insert('EventC')
                         insert('EventD')
                         insert('EventSp')
                         insert('GemsFarming')
 
         # Remove campaign_main from event list
-        for task in ['Event', 'Event2', 'EventAb', 'EventB', 'EventCd', 'EventD', 'EventSp', 'Raid', 'RaidDaily', 'WarArchives']:
+        for task in ['Event', 'Event2', 'EventA', 'EventB', 'EventC', 'EventD', 'EventSp', 'Raid', 'RaidDaily', 'WarArchives']:
             options = deep_get(self.args, keys=f'{task}.Campaign.Event.option')
             options = [option for option in options if option != 'campaign_main']
             deep_set(self.args, keys=f'{task}.Campaign.Event.option', value=options)
@@ -487,7 +487,7 @@ class ConfigUpdater:
         # Update to latest event
         server = to_server(deep_get(new, 'Alas.Emulator.PackageName', 'cn'))
         if not is_template:
-            for task in ['Event', 'Event2', 'EventAb', 'EventB', 'EventCd', 'EventD', 'EventSp', 'Raid', 'RaidDaily']:
+            for task in ['Event', 'Event2', 'EventA', 'EventB', 'EventC', 'EventD', 'EventSp', 'Raid', 'RaidDaily']:
                 deep_set(new,
                          keys=f'{task}.Campaign.Event',
                          value=deep_get(self.args, f'{task}.Campaign.Event.{server}'))
