+++ module/config/i18n/zh-CN.json
@@ -78,16 +78,16 @@
       "name": "活动图-2",
       "help": ""
     },
-    "EventAb": {
-      "name": "活动每日AB图",
+    "EventA": {
+      "name": "活动每日A图",
       "help": ""
     },
     "EventB": {
       "name": "活动每日B图",
       "help": ""
     },
-    "EventCd": {
-      "name": "活动每日CD图",
+    "EventC": {
+      "name": "活动每日C图",
       "help": ""
     },
     "EventD": {
@@ -947,7 +947,7 @@
     },
     "VanguardChange": {
       "name": "更换前排",
-      "help": "换前排通过找一艘150心情等级100的白鹰白皮驱逐完成，所以尽量保证有足够多的驱逐在后排休息。\n不启用将不更换前排"
+      "help": "换前排通过找一艘150心情等级100的白鹰白皮驱逐完成，所以尽量保证有足够多的驱逐在后排休息。\n国服以外则为150心情等级100的白鹰白船驱逐。\n不启用将不更换前排"
     },
     "VanguardEquipChange": {
       "name": "更换前排的装备",
@@ -981,9 +981,9 @@
       "help": "当本地时间超过给定数值后，禁用所有活动任务，'2020-01-01 00:00:00' 表示不限制，触发一次后时间将被重置\n设置后能防止活动结束后忘记关闭任务的事情发生"
     }
   },
-  "EventAb": {
+  "EventA": {
     "_info": {
-      "name": "活动每日AB图",
+      "name": "活动每日A图",
       "help": ""
     },
     "StageFilter": {
@@ -998,7 +998,7 @@
   "EventB": {
     "_info": {
       "name": "活动每日B图",
-      "help": "由于3油低耗队伍在B2.B3等地图会需要空弹打王，容易翻车\n故制作一个从每日AB分离的单独模块，来处理地图B2.B3"
+      "help": ""
     },
     "StageFilter": {
       "name": "活动每日过滤器",
@@ -1009,9 +1009,9 @@
       "help": "自动更新的数值，将跳过过滤器中已完成的关卡，过0点重置"
     }
   },
-  "EventCd": {
+  "EventC": {
     "_info": {
-      "name": "活动每日CD图",
+      "name": "活动每日C图",
       "help": ""
     },
     "StageFilter": {
@@ -1026,7 +1026,7 @@
   "EventD": {
     "_info": {
       "name": "活动每日D图",
-      "help": "由于CD图锁船种的要求不同，故制作一个从每日CD分离的单独模块来处理"
+      "help": ""
     },
     "StageFilter": {
       "name": "活动每日过滤器",
