+++ module/config/i18n/zh-TW.json
@@ -78,16 +78,16 @@
       "name": "活動圖-2",
       "help": ""
     },
-    "EventAb": {
-      "name": "活動每日AB圖",
+    "EventA": {
+      "name": "活動每日A圖",
       "help": ""
     },
     "EventB": {
       "name": "活動每日B圖",
       "help": ""
     },
-    "EventCd": {
-      "name": "活動每日CD圖",
+    "EventC": {
+      "name": "活動每日C圖",
       "help": ""
     },
     "EventD": {
@@ -947,7 +947,7 @@
     },
     "VanguardChange": {
       "name": "更換前排",
-      "help": "換前排通過找一艘150心情等級100的白鷹白船驅逐完成，所以盡量保證有足夠多的驅逐在後排休息。 \n不啟用將不更換前排"
+      "help": "換前排通過找一艘150心情等級100的白鷹白船驅逐完成，所以盡量保證有足夠多的驅逐在後排休息。\n國服以外則為150心情等級100的白鷹白船驅逐。 \n不啟用將不更換前排"
     },
     "VanguardEquipChange": {
       "name": "更換前排的裝備",
@@ -981,9 +981,9 @@
       "help": "當本地時間超過給定數值後，禁用所有活動任務，'2020-01-01 00:00:00' 表示不限制，觸發一次後時間將被重置\n設置後能防止活動結束後忘記關閉任務的事情發生"
     }
   },
-  "EventAb": {
+  "EventA": {
     "_info": {
-      "name": "活動每日AB圖",
+      "name": "活動每日A圖",
       "help": ""
     },
     "StageFilter": {
@@ -997,8 +997,8 @@
   },
   "EventB": {
     "_info": {
-      "name": "活動每日AB圖",
-      "help": "由於3油低耗隊伍在B2.B3等地圖會需要空彈打王，容易翻車\n故製作一個從每日AB分離的單獨模塊，來處理地圖B2.B3"
+      "name": "活動每日B圖",
+      "help": ""
     },
     "StageFilter": {
       "name": "活動每日過濾器",
@@ -1009,9 +1009,9 @@
       "help": "自動更新的數值，將跳過過濾器中已完成的地圖，過0點重置"
     }
   },
-  "EventCd": {
+  "EventC": {
     "_info": {
-      "name": "活動每日CD圖",
+      "name": "活動每日C圖",
       "help": ""
     },
     "StageFilter": {
@@ -1026,7 +1026,7 @@
   "EventD": {
     "_info": {
       "name": "活動每日D圖",
-      "help": "由於CD圖鎖船種的要求不同，故製作一個從每日CD分離的單獨模塊來處理"
+      "help": ""
     },
     "StageFilter": {
       "name": "活動每日過濾器",
